refactor(interactions): migrate raider buttons + user select to registry
