+++ src/events/interactionCreate.ts
@@ -10,12 +10,8 @@ import {
 import type { BotClient, TrialRow } from '../types/index.js';
 import { config } from '../config.js';
 import { logger } from '../services/logger.js';
-import { updateRaiderDiscordUser } from '../functions/raids/updateRaiderDiscordUser.js';
-import { ignoreCharacter } from '../functions/raids/ignoreCharacter.js';
-import { sendAlertForRaidersWithNoUser } from '../functions/raids/sendAlertForRaidersWithNoUser.js';
 import { audit, alertOfficers } from '../services/auditLog.js';
 import { getDatabase } from '../database/db.js';
-import type { RaiderRow } from '../types/index.js';
 import { startApplication } from '../functions/applications/startApplication.js';
 import { submitApplication } from '../functions/applications/submitApplication.js';
 import {
@@ -74,81 +70,6 @@ export default {
       if (await dispatch(buttonHandlers, 'button', interaction, customId)) return;
 
       try {
-        // raider:confirm_link:{characterName}:{userId}
-        if (customId.startsWith('raider:confirm_link:')) {
-          const parts = customId.split(':');
-          const characterName = parts[2];
-          const userId = parts[3];
-
-          const success = await updateRaiderDiscordUser(
-            interaction.client,
-            characterName,
-            userId,
-          );
-
-          if (success) {
-            await audit(interaction.user, 'confirmed raider link', `${characterName} -> <@${userId}>`);
-            await interaction.update({
-              content: `Linked **${characterName}** to <@${userId}>!`,
-              components: [],
-            });
-          } else {
-            await interaction.reply({
-              content: `Failed to link **${characterName}**.`,
-              flags: MessageFlags.Ephemeral,
-            });
-          }
-        }
-
-        // raider:reject_link:{characterName}
-        if (customId.startsWith('raider:reject_link:')) {
-          const characterName = customId.split(':')[2];
-
-          try {
-            await interaction.message.delete();
-          } catch {
-            // Message may already be deleted
-          }
-
-          // Post standard missing user alert (unmatched style)
-          const db = getDatabase();
-          const raider = db
-            .prepare('SELECT * FROM raiders WHERE character_name = ?')
-            .get(characterName) as RaiderRow | undefined;
-
-          if (raider) {
-            // Clear the old message_id since we deleted it
-            db.prepare('UPDATE raiders SET message_id = NULL WHERE character_name = ?').run(characterName);
-            await sendAlertForRaidersWithNoUser(interaction.client, [raider], []);
-          }
-        }
-
-        // raider:ignore:{characterName}
-        if (customId.startsWith('raider:ignore:')) {
-          const characterName = customId.split(':')[2];
-          const success = ignoreCharacter(characterName);
-
-          if (success) {
-            await audit(interaction.user, 'ignored character via button', characterName);
-
-            try {
-              await interaction.message.delete();
-            } catch {
-              // Message may already be deleted
-            }
-
-            await interaction.reply({
-              content: `Ignored **${characterName}** and removed from raiders.`,
-              flags: MessageFlags.Ephemeral,
-            });
-          } else {
-            await interaction.reply({
-              content: `Failed to ignore **${characterName}**.`,
-              flags: MessageFlags.Ephemeral,
-            });
-          }
-        }
-
         // application:apply - "Apply Now" button
         if (customId === 'application:apply') {
           const success = await startApplication(interaction.user);
@@ -412,52 +333,7 @@ export default {
     // User select menu handlers
     if (interaction.isUserSelectMenu()) {
       const customId = interaction.customId;
-
       if (await dispatch(userSelectHandlers, 'select', interaction, customId)) return;
-
-      try {
-        // raider:select_user:{characterName}
-        if (customId.startsWith('raider:select_user:')) {
-          const characterName = customId.split(':')[2];
-          const selectedUserId = interaction.values[0];
-
-          const success = await updateRaiderDiscordUser(
-            interaction.client,
-            characterName,
-            selectedUserId,
-          );
-
-          if (success) {
-            await audit(interaction.user, 'linked raider via select', `${characterName} -> <@${selectedUserId}>`);
-
-            try {
-              await interaction.message.delete();
-            } catch {
-              // Message may already be deleted
-            }
-
-            await interaction.reply({
-              content: `Linked **${characterName}** to <@${selectedUserId}>.`,
-              flags: MessageFlags.Ephemeral,
-            });
-          } else {
-            await interaction.reply({
-              content: `Failed to link **${characterName}**.`,
-              flags: MessageFlags.Ephemeral,
-            });
-          }
-        }
-      } catch (error) {
-        const err = error instanceof Error ? error : new Error(String(error));
-        logger.error('interaction', `Select menu handler failed (${customId}): ${err.message}`, err);
-
-        const reply = { content: 'An error occurred handling this selection.', flags: MessageFlags.Ephemeral } as const;
-        if (interaction.replied || interaction.deferred) {
-          await interaction.followUp(reply).catch(() => {});
-        } else {
-          await interaction.reply(reply).catch(() => {});
-        }
-      }
     }
 
     // Modal submit handlers
